+++ stac_fastapi/sfeos_helpers/stac_fastapi/sfeos_helpers/mappings.py
@@ -160,9 +160,7 @@ ES_COLLECTIONS_MAPPINGS = {
     "dynamic_templates": ES_MAPPINGS_DYNAMIC_TEMPLATES,
     "properties": {
         "id": {"type": "keyword"},
-        "title": {"type": "text"},
-        "description": {"type": "text"},
-        "bbox_shape": {"type": "geo_shape"},
+        "extent.spatial.bbox": {"type": "long"},
         "extent.temporal.interval": {
             "type": "date",
             "format": "strict_date_optional_time||epoch_millis",